+++ packages/cli/package.json
@@ -47,8 +47,8 @@
     "test:tokens-create-and-build-config": "yarn test:tokens-create-config && yarn test:tokens-build",
     "test": "yarn test:tokens-create-and-build-options && yarn test:tokens-create-and-build-config",
     "internal:tokens-create": "yarn designsystemet tokens create --config ./internal.config.json",
-    "update:template": "tsx ./src/tokens/template.ts",
-    "update:design-tokens": "yarn internal:tokens-create && tsx ./src/scripts/copy-internal-tokens.ts"
+    "update:template": "tsx ./src/scripts/update-template.ts",
+    "update:design-tokens": "yarn internal:tokens-create && tsx ./src/scripts/update-design-tokens.ts"
   },
   "dependencies": {
     "@commander-js/extra-typings": "^13.1.0",
