+++ packages/cli/src/tokens/template/design-tokens/primitives/globals.json
