+++ packages/cli/src/tokens/template/design-tokens/primitives/modes/size/large.json
