+++ packages/cli/src/tokens/template/design-tokens/primitives/modes/size/medium.json
