+++ packages/cli/src/tokens/template/design-tokens/primitives/modes/typography/size/medium.json
