+++ packages/cli/src/tokens/template/design-tokens/primitives/modes/typography/size/small.json
