+++ packages/cli/src/tokens/template/design-tokens/semantic/color-base-template.json
