+++ packages/cli/src/tokens/template/design-tokens/semantic/color-template.json
