+++ packages/cli/src/tokens/template/design-tokens/semantic/modes/category-color/color-template.json
