+++ packages/cli/src/tokens/template/design-tokens/semantic/style.json
