+++ packages/cli/src/tokens/template/design-tokens/themes/theme-base-template.json
