+++ packages/cli/src/tokens/template/design-tokens/themes/theme-template.json
